ll6

+++ sde-sheet/07_linkedlistnarray/6_maxonsecutiveones.cpp
@@ -13,11 +13,11 @@
 
 #include <bits/stdc++.h>
 using namespace std;
-int findMaxConsecutiveOnes(vector < int > & nums) {
+int findMaxConsecutiveOnes(vector<int> &nums) {
 	int cnt = 0;
 	int maxi = 0;
-	for (int i = 0; i < (int)nums.size(); i++) {
-		if (nums[i] == 1)
+	for(int i=0; i<nums.size(); i++) {
+		if(nums[i] == 1)
 			cnt++;
 		else
 			cnt = 0;
